update currency to integer and use make:cast to cast the column to money
https://filamentphp.com/docs/3.x/panels/getting-started#casting-the-price-to-an-integer

+++ database/migrations/2020_03_12_155802_create_items_table.php
@@ -20,7 +20,7 @@ class CreateItemsTable extends Migration
             $table->string('description');
             $table->decimal('duration_rate')->default(0);
             $table->string('duration')->nullable();
-            $table->decimal('price')->nullable()->default(0);
+            $table->integer('price')->nullable()->default(0);
             $table->boolean('obligatory')->default(false);
 
             $table->unsignedInteger('order')->default(0);
